+++ include/geos/geom/GeometryFactory.h
@@ -282,6 +282,15 @@ public:
     std::unique_ptr<Geometry> createEmpty(int dimension) const;
 
     /**
+    * Creates an empty atomic geometry of the given type.
+    * @param typeId the desired GeometryTypeId
+    * @return an empty atomic geometry of given dimension
+    */
+    std::unique_ptr<Geometry> createEmpty(GeometryTypeId typeId) const;
+
+    std::unique_ptr<Geometry> createMulti(std::unique_ptr<Geometry> && geom) const;
+
+    /**
      *  Build an appropriate <code>Geometry</code>, <code>MultiGeometry</code>, or
      *  <code>GeometryCollection</code> to contain the <code>Geometry</code>s in
      *  it.