+++ NEWS.md
@@ -52,6 +52,7 @@ xxxx-xx-xx
   - MaximumInscribedCircle: Fix infinite loop with non-finite coordinates (GH-843, Dan Baston)
   - DistanceOp: Fix crash on collection containing empty point (GH-842, Dan Baston)
   - Improve OffsetCurve behaviour and add Joined mode (JTS-956, Martin Davis)
+  - GeometryPrecisionReducer preserves input collection types (GH-846, Paul Ramsey)
 
 ## Changes in 3.11.0
 2022-07-01